(~) UserPreferences-Datenklasse in eigene Datei extrahiert (500-Zeilen-Limit)
